refactor: replace custom Kratos filter with Spring OpaqueTokenIntrospector
Addresses review feedback to lean on Spring Security instead of custom auth
plumbing. The Kratos /sessions/whoami call now runs inside an
OpaqueTokenIntrospector; Spring's standard BearerTokenAuthenticationFilter
handles header parsing, 401 + WWW-Authenticate, and SecurityContext setup.

Downstream contract preserved: principal.name == user.code, authorities
unchanged. Behavior change: 5xx-from-Kratos now surfaces as 401 (was 503),
documented in ADR 0001.
