+++ docs/adr/0001-mobile-auth-via-kratos-session-introspection.md
@@ -29,10 +29,11 @@ For backend acceptance of the resulting Kratos session token, three options were
 
 - The KMP app opens `https://auth.flock.community/...` in an in-app authentication session (`ASWebAuthenticationSession` on iOS, Chrome Custom Tabs on Android), completes the Kratos browser flow, captures the returned Kratos session token via a deep-link callback, and stores it in Keychain / EncryptedSharedPreferences.
 - The app sends `Authorization: Bearer <kratos-session-token>` on every API request.
-- A new `KratosSessionFilter` in `workday-application/.../config/` reads the bearer token, calls Kratos `sessions/whoami` (with explicit connect/read timeouts so a slow Kratos cannot pin Tomcat threads, and a bounded LRU cache — positive ~60s, negative ~10s — to absorb token-spray without hammering Kratos), and resolves the Kratos identity to a workday `User` via `userService.findByEmail(traits.email)` — creating the `User` if missing. It then populates the Spring `SecurityContext` with a `UsernamePasswordAuthenticationToken(user.code, null, user.getGrantedAuthority())`, matching the pattern used by the existing `UserKeyTokenFilter`.
+- A new `KratosOpaqueTokenIntrospector` in `workday-application/.../config/` is registered via Spring Security's standard `oauth2ResourceServer().opaqueToken().introspector(...)`. The introspector calls Kratos `sessions/whoami` (with explicit connect/read timeouts so a slow Kratos cannot pin Tomcat threads, and a bounded LRU cache — positive ~60s, negative ~10s — to absorb token-spray without hammering Kratos), and resolves the Kratos identity to a workday `User` via `userService.findByEmail(traits.email)` — creating the `User` if missing. It then returns an `OAuth2AuthenticatedPrincipal` whose `name == user.code`, so downstream `authentication.name`-based lookups (e.g. `personService.findByUserCode`) keep working unchanged. Spring's stock `BearerTokenAuthenticationFilter` handles `Authorization: Bearer` parsing, 401 + `WWW-Authenticate` on invalid tokens, and `SecurityContext` population — the only project-specific code is the introspector itself. Note: Kratos `/sessions/whoami` is not RFC 7662, so this uses the `OpaqueTokenIntrospector` abstraction without the standard underneath it; the trade-off is deliberate (reviewer feedback: prefer Spring Security's library shape over a custom `GenericFilterBean`).
 - **The filter does not touch `UserAccountOauth`.** Its mapping (Google `sub` → user) remains the responsibility of the legacy `UserSecurityService.googleLogin` path. A mobile-first user will get a `User` row but no `UserAccountOauth` row; if they later log in via web, the existing `googleLogin` code finds the user by email and backfills the oauth row.
 - Rationale for avoiding `UserSecurityOauth2`: it wraps an `OidcIdToken`, which Kratos `sessions/whoami` does not provide. Constructing a synthetic token would be misleading.
-- The existing `WebSecurityConfig` filter chain remains; the new filter is added alongside `UserKeyTokenFilter`.
+- The existing `WebSecurityConfig` filter chain remains; `UserKeyTokenFilter` keeps handling `Authorization: TOKEN <api-key>` and the OAuth2 resource-server filters handle `Authorization: Bearer <kratos-token>`. The two prefixes are disjoint.
+- Operational simplification vs an earlier iteration: Kratos-unreachable failures surface as `OAuth2IntrospectionException`, which Spring renders as the standard 401 rather than the 503 the first filter-based draft returned. Mobile clients see a uniform 401 on any auth failure and re-auth; a custom `AuthenticationEntryPoint` to recover the 503 distinction was considered and rejected to keep the integration aligned with off-the-shelf Spring Security defaults.
 
 **Post-MVP (option iii):** migrate to Oathkeeper-fronted bearer-token auth with the workday backend validating the Oathkeeper-injected JWT against `terraform/oathkeeper/jwks/jwks.json`. This removes the per-request `sessions/whoami` round-trip and aligns mobile + web on a single trust model. Tracked as a follow-up; not in MVP scope.
